+++ app/Http/Controllers/Home/Forum/PostController.php
@@ -18,6 +18,12 @@ class PostController extends Controller
 
     public function like(Forum $forum,Post $post)
     {
+    	// Check if block
+    	if(auth('user')->user()->isForumBlock($forum))
+	    {
+	    	return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->withErrors(['شما توسط مدیر انجمن محدود شده اید و امکان انجام این درخواست وجود ندارد']);
+	    }
+
         // Detach dislike
         $post->usersDislike()->detach(auth('user')->id());
 
@@ -30,18 +36,41 @@ class PostController extends Controller
 
         return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->with('success',['درخواست شما با موفقیت انجام شد']);
     }
-    public function dislike(Forum $forum,Post $post)
-    {
-        // Detach like
-        $post->usersLike()->detach(auth('user')->id());
+	public function dislike(Forum $forum,Post $post)
+	{
+		// Check if block
+		if(auth('user')->user()->isForumBlock($forum))
+		{
+			return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->withErrors(['شما توسط مدیر انجمن محدود شده اید و امکان انجام این درخواست وجود ندارد']);
+		}
 
-        if($post->isUserDislike(auth('user')->user()))
-        {
-            $post->usersDislike()->detach(auth('user')->id());
-        }else{
-            $post->usersDislike()->attach(auth('user')->id());
-        }
+		// Detach like
+		$post->usersLike()->detach(auth('user')->id());
 
-        return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->with('success',['درخواست شما با موفقیت انجام شد']);
-    }
+		if($post->isUserDislike(auth('user')->user()))
+		{
+			$post->usersDislike()->detach(auth('user')->id());
+		}else{
+			$post->usersDislike()->attach(auth('user')->id());
+		}
+
+		return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->with('success',['درخواست شما با موفقیت انجام شد']);
+	}
+	public function save(Forum $forum,Post $post)
+	{
+		// Check if block
+		if(auth('user')->user()->isForumBlock($forum))
+		{
+			return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->withErrors(['شما توسط مدیر انجمن محدود شده اید و امکان انجام این درخواست وجود ندارد']);
+		}
+
+		if($post->isUserSave(auth('user')->user()))
+		{
+			$post->usersSave()->detach(auth('user')->id());
+		}else{
+			$post->usersSave()->attach(auth('user')->id());
+		}
+
+		return redirect()->route('home.forum.post.show',[$forum->slug,$post->id])->with('success',['درخواست شما با موفقیت انجام شد']);
+	}
 }
